requirements.txt
# Agente 1 - Scraper y Data Engineer
apify-client>=1.9.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0

# Grafo LangGraph v1 (compatible con LangChain v1)
langgraph>=1.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# Configuración centralizada
pydantic-settings>=2.0.0

# Rate limiting
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
